+++ CHANGELOG.md
@@ -28,8 +28,9 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ------
 ## [v10.0.0](https://github.com/asfadmin/Discovery-asf_search/compare/v9.0.9...v10.0.0)
 ### Added
-- adds `asf-enumeration` package as dependency
+- adds `asf-enumeration` package as optional dependency (installable via `pip install asf-search[asf-enumeration]`)
     - Enables SLC baseline stacking based on `ARIA S1 GUNW` frame
+       - `ARIAS1GUNWProduct.stack()` returns stack of SLCs over ARIA Frame
        - to use with `stack_from_id()`, pass `ASFSearchOptions` object to `opts` with `dataset` set to `constants.DATASET.ARIA_S1_GUNW`
 
 ### Changed